Refactor: Domain split into readDomain and writeDomain
