+++ src/main/kotlin/Main.kt
@@ -1,6 +1,7 @@
-import domain.Domain
 import adapters.HttpApi
 import adapters.TodoListFileRepo
+import domain.ReadDomain
+import domain.WriteDomain
 import ports.TodoListRepo
 import org.http4k.core.HttpHandler
 import org.http4k.core.then
@@ -10,9 +11,11 @@ import org.http4k.server.asServer
 
 fun main() {
     val todoListRepo: TodoListRepo = TodoListFileRepo("./src/resources/todo_list.json")
-    val domain = Domain(todoListRepo)
+    val readDomain = ReadDomain(todoListRepo)
+    val writeDomain = WriteDomain(todoListRepo, readDomain)
+    val httpApi = HttpApi(readDomain, writeDomain)
 
-    val printingApp: HttpHandler = DebuggingFilters.PrintRequest().then(HttpApi(domain).app)
+    val printingApp: HttpHandler = DebuggingFilters.PrintRequest().then(httpApi.app)
 
     val server = printingApp.asServer(SunHttp(3000)).start()
