+++ src/main/kotlin/adapters/HttpAPI.kt
@@ -1,32 +1,32 @@
 package adapters
 
-import domain.Domain
 import domain.models.TodoItem
 import org.http4k.core.Method.GET
 import org.http4k.core.Method.POST
 import org.http4k.core.Method.PATCH
-import org.http4k.core.Method.DELETE
 import org.http4k.core.Status.Companion.OK
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import domain.ReadDomain
+import domain.WriteDomain
 import org.http4k.core.*
 import org.http4k.core.Status.Companion.CREATED
 import org.http4k.routing.*
 
-class HttpApi(domain: Domain) {
+class HttpApi(readDomain: ReadDomain, writeDomain: WriteDomain) {
 
     val app: HttpHandler = routes(
         "/todos" bind GET to { request: Request ->
             val todoStatus: String = request.query("status") ?: ""
 
             if (todoStatus == "") {
-                val todoList: List<TodoItem> = domain.getTodoList()
+                val todoList: List<TodoItem> = readDomain.getTodoList()
                 val toDoListAsJsonString: String = mapper.writeValueAsString(todoList)
                 Response(OK)
                     .body(toDoListAsJsonString)
                     .header("content-type","application/json")
             } else {
-                val todoListFilteredByStatus: List<TodoItem> = domain.getItemsByStatus(todoStatus)
+                val todoListFilteredByStatus: List<TodoItem> = readDomain.getItemsByStatus(todoStatus)
                 val todoListFilteredByStatusAsJSONString: String = mapper.writeValueAsString(todoListFilteredByStatus)
                 Response(OK)
                     .body(todoListFilteredByStatusAsJSONString)
@@ -37,7 +37,7 @@ class HttpApi(domain: Domain) {
         "/todos" bind POST to {request: Request ->
             val newTodoData: String  = request.bodyString()  //todo handle errors
             val newTodoName: String = mapper.readTree(newTodoData).get("name").asText()
-            val newTodo = domain.addTodo(newTodoName)
+            val newTodo = writeDomain.addTodo(newTodoName)
             val newTodoId = newTodo.id
             val newTodoURL = "http://localhost:3000/todos/${newTodoId}"
             val newTodoAsJsonString = mapper.writeValueAsString(newTodo)
@@ -55,18 +55,18 @@ class HttpApi(domain: Domain) {
             val todoStatusUpdate: String? = mapper.readTree(todoDataToUpdate).get("status")?.asText()
 
             if (todoNameUpdate != null) {
-                domain.updateTodoName(todoId, todoNameUpdate)
+                writeDomain.updateTodoName(todoId, todoNameUpdate)
             }
 
             if (todoStatusUpdate != null && todoStatusUpdate == "DONE") {
-                domain.markTodoAsDone(todoId)
+                writeDomain.markTodoAsDone(todoId)
             }
 
             if (todoStatusUpdate != null && todoStatusUpdate == "NOT_DONE") {
-                domain.markTodoAsNotDone(todoId)
+                writeDomain.markTodoAsNotDone(todoId)
             }
 
-            val updatedTodo: List<TodoItem> = domain.getTodoList(todoId)
+            val updatedTodo: List<TodoItem> = readDomain.getTodoList(todoId)
             val updatedTodoAsJson = mapper.writeValueAsString(updatedTodo)
 
                 Response(OK)
@@ -76,18 +76,12 @@ class HttpApi(domain: Domain) {
 
         "/todos/{todoId}" bind GET to {request: Request ->
             val todoId: String = request.path("todoId")!! // handle errors if id is not valid
-            val todoItem: List<TodoItem> = domain.getTodoList(todoId)
+            val todoItem: List<TodoItem> = readDomain.getTodoList(todoId)
             val toDoListAsJsonString: String = mapper.writeValueAsString(todoItem)
             Response(OK)
                 .body(toDoListAsJsonString)
                 .header("content-type", "application/json")
         },
-
-        "/todos/{todoId}" bind DELETE to {request: Request ->
-            val todoId: String = request.path("todoId")!!
-            val todoDeletionConfirmation = domain.deleteTodo(todoId)
-            Response(OK).body(todoDeletionConfirmation)
-        }
 
     )
